Project Rock Paper Scissors
This is my first Javascript program built from scratch. For now, the game is going to be played in the browser console. I started by creating a function called computerPlay that will randomly return either ‘Rock’, ‘Paper’ or ‘Scissors’. After which I made another function called playRound that plays a single round of the game.

Finally, I made another function called game that plays a 5 round of the game by making use of for loops.